feat: refactor go to payment button
