+++ src/components/sign-up/PriceCalculation.tsx
@@ -1,56 +1,15 @@
-import { useMemo } from "react";
-import dayjs from "dayjs";
-import { CATEGORIES } from "../../constants/categories";
 import { H3 } from "../../styled-components/Types";
 import styled from "styled-components";
 import { Particpant } from "../../types/Participant";
-import { Button } from "../core";
-import {
-  allParticipantsAreValid,
-  participantIsValid,
-} from "../../schema/Participant";
+
+import { usePricing } from "../../hooks/usePricing";
 
 type PriceCalculationProps = {
   participants: Particpant[];
 };
 
 export const PriceCalculation = ({ participants }: PriceCalculationProps) => {
-  const pricing = useMemo(() => {
-    const selectedCategories = participants
-      .filter((participant) => participantIsValid(participant).valid)
-      .map((participant) => {
-        const age = dayjs().diff(participant.dateOfBirth, "years");
-
-        const category = Object.values(CATEGORIES).find((categoryInfo) => {
-          return age >= categoryInfo.minAge && age <= categoryInfo.maxAge;
-        });
-
-        return category;
-      });
-
-    const adultsAmount = selectedCategories.filter(
-      (cat) => !cat?.isChild
-    ).length;
-    const adultPrice = CATEGORIES.adult.price;
-
-    const childrenAmount = selectedCategories.filter(
-      (cat) => cat?.isChild
-    ).length;
-    const childPrice = CATEGORIES.child1.price;
-
-    return {
-      adults: {
-        amount: adultsAmount,
-        pricePP: CATEGORIES.adult.price,
-        priceTotal: adultsAmount * adultPrice,
-      },
-      children: {
-        amount: childrenAmount,
-        pricePP: CATEGORIES.child1.price,
-        priceTotal: childrenAmount * childPrice,
-      },
-    };
-  }, [participants]);
+  const { pricing } = usePricing(participants);
 
   return (
     <StContainer>
@@ -82,22 +41,6 @@ export const PriceCalculation = ({ participants }: PriceCalculationProps) => {
           </StPrice>
         </StPriceDetailContainer>
       </StPriceRow>
-
-      <StPaymentButtonContainer>
-        <Button
-          disabled={!allParticipantsAreValid(participants)}
-          onClick={() => {
-            alert(
-              "Open betaling voor" +
-                pricing.children.priceTotal +
-                pricing.adults.priceTotal +
-                "euro"
-            );
-          }}
-          text={"Ga naar betaling"}
-          size={"medium"}
-        />
-      </StPaymentButtonContainer>
     </StContainer>
   );
 };
@@ -130,10 +73,4 @@ const StLine = styled.div`
   height: 1px;
   width: 100%;
   background-color: black;
-`;
-
-const StPaymentButtonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  margin-top: 48px;
 `;
